refactor: 패키지 이동 및  api 실행 제어
