+++ backend/src/main/java/com/carffeine/carffeine/admin/service/AdminStationService.java
@@ -7,8 +7,8 @@ import com.carffeine.carffeine.member.domain.Member;
 import com.carffeine.carffeine.member.domain.MemberRepository;
 import com.carffeine.carffeine.station.domain.station.Station;
 import com.carffeine.carffeine.station.domain.station.StationRepository;
-import com.carffeine.carffeine.station.exception.station.StationException;
-import com.carffeine.carffeine.station.exception.station.StationExceptionType;
+import com.carffeine.carffeine.station.exception.StationException;
+import com.carffeine.carffeine.station.exception.StationExceptionType;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
