+++ backend/src/main/java/com/carffeine/carffeine/station/service/city/CityRequester.java
@@ -1,0 +1,10 @@
+package com.carffeine.carffeine.station.service.city;
+
+import com.carffeine.carffeine.station.domain.city.City;
+
+import java.util.List;
+
+public interface CityRequester {
+
+    List<City> requestCities();
+}
